CProxy destroy-check added if a new subprocess should be started. Added missing documentation for methods. Added 'destroyed?' method to check destruction.

+++ lib/ruby_process_cproxy.rb
@@ -28,7 +28,7 @@ class Ruby_process::Cproxy
       @@instances += 1
       
       #Check if the sub-process is alive.
-      if @@subproc and !@@subproc.alive?
+      if @@subproc and !@@subproc.alive? and !@@subproc.destroyed?
         @@subproc.destroy
         @@subproc = nil
       end
